날짜: 2020.06.16. 작성자: 신현준 내용: 버튼 CSS 작업, 게시판 수정 기능 수정

+++ NaeMentor/src/main/webapp/WEB-INF/views/Notiquestion/Notification_boardDetail.jsp
@@ -22,16 +22,18 @@
 		</div>
 			<div style="text-align: center; margin-top: 20px;">
 			<c:if test="${userinfo.auth eq 'ROLE_A'}">	
-			<input type="button" name="btn" onclick="notidel()" value="삭제" >
+			<input type="button" class="myButton" name="btn" onclick="notidel()" value="삭제" >
 			</c:if>
-			<input type="button" name="btn" onclick="notilist()" value="목록보기" >
+			<input type="button" class="myButton" name="btn" onclick="notilist()" value="목록보기" >
 			<c:if test="${userinfo.auth eq 'ROLE_A'}">
-			<input type="button" name="btn" onclick="notimod()" value="수정">
+			<input type="button" class="myButton" name="btn" onclick="notimod()" value="수정">
 			</c:if>
 		</div>
 		<script type="text/javascript">
 			function notidel(){
+				if (confirm("정말로 삭제하시겠습니까?")){
 				location.href="./Notification_boardDelete.do?adminseq="+${dto.adminseq};
+				}
 			}
 			function notilist(){
 				location.href="./Notification_board.do";
